move dbc related code to mainwin
