+++ tools/cabana/messageswidget.h
@@ -1,32 +1,9 @@
 #pragma once
 
 #include <QAbstractTableModel>
-#include <QComboBox>
-#include <QDialog>
-#include <QJsonDocument>
 #include <QTableView>
-#include <QTextEdit>
 
 #include "tools/cabana/canmessages.h"
-
-class LoadDBCDialog : public QDialog {
-  Q_OBJECT
-
-public:
-  LoadDBCDialog(QWidget *parent);
-  QTextEdit *dbc_edit;
-};
-
-class SaveDBCDialog : public QDialog {
-  Q_OBJECT
-
-public:
-  SaveDBCDialog(QWidget *parent);
-  void copytoClipboard();
-  void saveAs();
-  QTextEdit *dbc_edit;
-};
-
 class MessageListModel : public QAbstractTableModel {
 Q_OBJECT
 
@@ -58,18 +35,10 @@ class MessagesWidget : public QWidget {
 public:
   MessagesWidget(QWidget *parent);
 
-public slots:
-  void loadDBCFromName(const QString &name);
-  void loadDBCFromFingerprint();
-  void loadDBCFromPaste();
-  void saveDBC();
-
 signals:
   void msgSelectionChanged(const QString &message_id);
 
 protected:
   QTableView *table_widget;
-  QComboBox *dbc_combo;
   MessageListModel *model;
-  QJsonDocument fingerprint_to_dbc;
 };
